Fix pt cut.

+++ PhysicsTools/PatExamples/test/analyzePatBJetTags_cfg.py
@@ -43,7 +43,7 @@ process.analyzeBJetTags = cms.EDAnalyzer("PatBJetTagAnalyzer",
 	jets = cms.InputTag("selectedLayer1Jets"),
 
 	#jet cuts
-	jetPtCut = cms.double(50),
+	jetPtCut = cms.double(30),
 	jetEtaCut = cms.double(2.4)
 )
